Add hello world generator
We want to be able to generate code based on custom templates.  This creates a
"hello world" generator that was created using the swagger-codegen jar's meta
utility.

The template is called typescript but the actual output is all from the
example generator that the swagger codegen library produces when you
tell it to create a custom class.

Issue #5 Add tooling for invoking code generation