Support dynamic package version
We need the package version of the generated Typescript file to change
for every distinct release. This means we have to combine two separate
semver values into one (the two being the version of the template, and
the version of the open API specification).

The logic outlined here should make sure the resulting commit will
always adhere to semver's rules around major changes -- in particular we
know that if EITHER version is still in 0.x.x then a patch change COULD
be a breaking change, and so we need to release the package as 0.x.x as
well in that case.

Another oddity of this commit is that we need to have .json AND
.mustache versions of the package / package lock files.  This is because
dependabot expects the `.json` variety whereas codegen requires
`.mustache` file names for the file to be treated as a template.

Issue #9 Populate dynamic version in package.json