Improve documentation
We had instructions for generating code but it was written assuming the
user was already aware of the repository context / had already copied
the code.  We want this README to be useful as a "first point" of entry,
so adding some steps for copying the code base should help.

Issue #14