Migrate to OpenAPI Generator
The Swagger Codegen project was forked a few years back and the majority
of the community moved over to the forked version [1].  This explains
why we were facing some odd issues and a lack of support for the Open
API specification 3.1

Moving to this more maintained version of the codegen project will
immediately result in some bug fixes (specifically, `object` will be
mapped to `object` instead of `any`).  It will also allow us to use 3.1
in our api specification.

Eventually this change may allow us to leverage community-defined
typescript templates instead of our own custom templates.

Issue #25

[1] https://github.com/OpenAPITools/openapi-generator/blob/master/docs/qna.md